Add Signal Plans and update Device Event Logging Configuration

+++ Atspm/Data/Interfaces/IPlanLayer.cs
@@ -25,8 +25,10 @@ namespace Utah.Udot.Atspm.Data.Interfaces
     public interface IPlanLayer : ILocationLayer
     {
         /// <summary>
-        /// Plan number as derrived from the event parameter on <see cref="IndianaEnumerations.CoordPatternChange"/> event
+        /// The identifier of the signal timing plan active during the aggregation period.  
+        /// Used to associate performance metrics with specific coordination patterns
+        /// or time‑of‑day schedules.
         /// </summary>
-        int PlanNumber { get; set; }
+        short PlanNumber { get; set; }
     }
 }
